+++ src/oop.cpp
@@ -2,19 +2,12 @@
 #include <algorithm>
 #include "class.hpp"
 #include "class_member.hpp"
-#include "memory.hpp"
 #include "thread.hpp"
-#include "method_handle.hpp"
-#include "string_pool.hpp"
 
 
 namespace RexVM {
 
     SpinLock Oop::monitorLock;
-
-    constexpr u2 TRACED_MASK = 0x8000;     //1000000000000000
-    constexpr u2 MIRROR_MASK = 0x4000;     //0100000000000000
-    constexpr u2 FINALIZED_MASK = 0x2000;  //0010000000000000
 
     Oop::Oop(Class *klass, size_t dataLength) :
             comClass(klass, dataLength),
@@ -245,70 +238,6 @@ namespace RexVM {
 
     InstanceClass *InstanceOop::getInstanceClass() const {
         return CAST_INSTANCE_CLASS(getClass());
-    }
-
-    MirOop::MirOop(InstanceClass *klass, voidPtr mirrorObj, MirrorObjectTypeEnum type) :
-        InstanceOop(klass, klass->instanceSlotCount),
-        mirror(mirrorObj, static_cast<u2>(type)) {
-        setFlags(getFlags() | MIRROR_MASK);
-    }
-
-    void MirOop::clearHolder() {
-        const auto type = getMirrorObjectType();
-        switch (type) {
-            case MirrorObjectTypeEnum::CLASS: {
-                const auto mirrorClass = getMirrorClass();
-                mirrorClass->mirrorBase.clear(this);
-                break;
-            }
-
-            case MirrorObjectTypeEnum::FIELD:
-            case MirrorObjectTypeEnum::METHOD:
-            case MirrorObjectTypeEnum::CONSTRUCTOR: {
-                const auto member = CAST_CLASS_MEMBER(mirror.getPtr());
-                member->mirrorBase.clear(this);
-                break;
-            }
-
-            case MirrorObjectTypeEnum::CONSTANT_POOL: {
-                const auto mirrorClass = CAST_INSTANCE_CLASS(getMirrorClass());
-                mirrorClass->constantPoolMirrorBase.clear(this);
-            }
-
-            default:
-                break;
-        }
-    }
-
-    MirOop::~MirOop() {
-        clearHolder();
-    }
-
-    MirrorObjectTypeEnum MirOop::getMirrorObjectType() const {
-        return static_cast<MirrorObjectTypeEnum>(mirror.getData());
-    }
-
-    Class *MirOop::getMirrorClass() const {
-        return CAST_CLASS(mirror.getPtr());
-    }
-
-    Method *MirOop::getMirrorMethod() const {
-        return CAST_METHOD(mirror.getPtr());
-    }
-
-    Field *MirOop::getMirrorField() const {
-        return CAST_FIELD(mirror.getPtr());
-    }
-
-    Method *MirOop::getMemberNameMethod() {
-        auto methodPtr = CAST_METHOD(mirror.getPtr());
-        if (methodPtr == nullptr) {
-            auto [klass, name, type, flags, kind, isStatic, descriptor]
-                    = methodHandleGetFieldFromMemberName(this);
-            methodPtr = klass->getMethod(name, descriptor, isStatic);
-            mirror.setPtr(methodPtr);
-        }
-        return methodPtr;
     }
 
     ArrayOop::ArrayOop(const OopTypeEnum type, ArrayClass *klass, const size_t dataLength) :
